project/requirements.txt
streamlit==1.44.1
numpy==1.26.4; python_version < "3.13"
numpy>=2.1,<2.3; python_version >= "3.13"
pandas==2.2.3
scikit-learn==1.5.2; python_version < "3.13"
scikit-learn>=1.6,<1.8; python_version >= "3.13"
spacy==3.7.5; python_version < "3.13"
spacy>=3.8,<3.9; python_version >= "3.13"
sentence-transformers==3.4.1
PyMuPDF==1.25.5
python-docx==1.1.2
pypdf>=5.0.0
openai>=1.52.0
reportlab>=4.2.2
python-dotenv>=1.0.1
